refactor(materializer): extract operation-index-patch subsystem (file #3)
Lift the full operation-index patching subsystem (incremental patch strategies +
the 4 lazy patched/appended Map subclasses + affected-entry collection, 16 fns /
4 classes) into operation-index-patch.ts. Self-contained; entry points
applyOperationIndexPatch, applyStableMarkedOperationIndexPatchInPlace,
collectAffectedTranscriptEntryIds. Net 74/74, check 0.
